Move solver-level types to src/solver_types.rs
Assertion, Quantifier, TermOption, ConstructorType, Polarity moved out of
egraphs/datastructures.rs into solver_types.rs. The egraphs/ module now
only contains egraph-internal types (DisequalTerm, Predecessor, CanonicalForm).
