+++ src/lib.rs
@@ -15,4 +15,5 @@ pub mod preprocess;
 mod proof;
 mod quantifiers;
 pub mod solver_state;
+pub mod solver_types;
 pub mod stats;
